[WIP] feat(iob-soc): Update IOBSOC submodule and configuration; Remove TESTER. - New IOBSOC submodule will function as a base for the SUT and Tester   system. The TESTER submodule is no longer required.   - Update `iob_soc_sut_setup.py` to remove most configuration common     with iob-soc, and add iob-soc parameters.   - Add `iob_soc_tester` folder with configuration specific for the     TESTER.
