+++ .gitmodules
@@ -13,9 +13,6 @@
 [submodule "submodules/REGFILEIF"]
 	path = submodules/REGFILEIF
 	url = git@github.com:IObundle/iob-regfileif.git
-[submodule "submodules/TESTER"]
-	path = submodules/TESTER
-	url = git@github.com:IObundle/iob-soc-tester.git
 [submodule "submodules/ETHERNET"]
 	path = submodules/ETHERNET
 	url = git@github.com:IObundle/iob-eth.git
